# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/accounting/report_link": {
            "post": {
                "description": "Метод по году и месяцу возвращает ссылку на скачивание отчёта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounting"
                ],
                "summary": "Получение ссылки с отчётом за месяц",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/avito2022_internal_app_api.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/avito2022_internal_app_api.ReportLink"
                        }
                    }
                }
            }
        },
        "/balance/get_balance": {
            "post": {
                "description": "Возвращает баланс пользователя по его id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Получение баланса пользователя",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.UserID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Balance"
                        }
                    }
                }
            }
        },
        "/balance/transfer_money": {
            "post": {
                "description": "Метод перевода денег от пользователя к пользователю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Перевод денег",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Transfer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/balance/up_balance": {
            "post": {
                "description": "Пополняет баланс пользователя или создаёт его при первом пополнении",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Пополнение или инициализация баланса",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.BalanceReplenishment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/balance/withdraw_money": {
            "post": {
                "description": "Метод снимает указанное количество средств со счёта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Снятие денег со счёта",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.DebitingMoney"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sales/reserve_money": {
            "post": {
                "description": "Резервирование средств с основного баланса на отдельном счете",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "sales"
                ],
                "summary": "Резервирование средств",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Reserve"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sales/return_money": {
            "post": {
                "description": "Разрезервирование средств и возвращение их на баланс пользователя при отмене операции",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "sales"
                ],
                "summary": "Возвращение средств",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Refund"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sales/revenue_confirmation": {
            "post": {
                "description": "Списывает из резерва деньги, добавляет данные в отчет для бухгалтерии",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "sales"
                ],
                "summary": "Метод признания выручки",
                "parameters": [
                    {
                        "description": "Входные параметры",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Confirmation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user_transaction/cost_sort": {
            "get": {
                "description": "Сортирует транзакции пользователя по стоимости и с использованием пагинации",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pagination"
                ],
                "summary": "Сортировка транзакций по стоимости c пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "сортировать по убыванию или возрастанию",
                        "name": "increase",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "курсор пагинации",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Page"
                        }
                    }
                }
            }
        },
        "/user_transaction/time_sort": {
            "get": {
                "description": "Сортирует транзакции пользователя по времени и с использованием пагинации",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pagination"
                ],
                "summary": "Сортировка транзакций по времени с пагинацией",
                "parameters": [
                    {
                        "type": "string",
                        "description": "сортировать по убыванию или возрастанию",
                        "name": "increase",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id пользователя",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "курсор пагинации",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_api.Page"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "avito2022_internal_app_api.Balance": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
        "avito2022_internal_app_api.BalanceReplenishment": {
            "type": "object",
            "properties": {
                "replenishment": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "avito2022_internal_app_api.Confirmation": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "avito2022_internal_app_api.DebitingMoney": {
            "type": "object",
            "properties": {
                "debit_cost": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "avito2022_internal_app_api.Page": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pagination.Report"
                    }
                },
                "next_cursor": {
                    "type": "string"
                },
                "prev_cursor": {
                    "type": "string"
                }
            }
        },
        "avito2022_internal_app_api.Refund": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "avito2022_internal_app_api.ReportLink": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                }
            }
        },
        "avito2022_internal_app_api.ReportRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "avito2022_internal_app_api.Reserve": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "avito2022_internal_app_api.Transfer": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "recipient_name": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                }
            }
        },
        "avito2022_internal_app_api.UserID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.Balance": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
        "internal_app_api.BalanceReplenishment": {
            "type": "object",
            "properties": {
                "replenishment": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.Confirmation": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.DebitingMoney": {
            "type": "object",
            "properties": {
                "debit_cost": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.Page": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pagination.Report"
                    }
                },
                "next_cursor": {
                    "type": "string"
                },
                "prev_cursor": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.Refund": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.ReportLink": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.ReportRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.Reserve": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "order_id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "service_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_app_api.Transfer": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "recipient_id": {
                    "type": "integer"
                },
                "recipient_name": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "sender_name": {
                    "type": "string"
                }
            }
        },
        "internal_app_api.UserID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "pagination.Report": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "operation_time": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        }
    }
}